+++ pyccel/ast/builtin_methods/list_methods.py
@@ -14,7 +14,10 @@ from pyccel.ast.datatypes import NativeVoid, NativeGeneric, NativeHomogeneousLis
 from pyccel.ast.internals import PyccelInternalFunction
 
 
-__all__ = ('ListAppend',)
+__all__ = ('ListAppend',
+           'ListClear',
+           'ListPop'
+           )
 
 
 class ListAppend(PyccelInternalFunction):
@@ -79,3 +82,90 @@ class ListAppend(PyccelInternalFunction):
         Get the argument which is passed to append().
         """
         return self._append_arg
+
+#==============================================================================
+class ListPop(PyccelInternalFunction) :
+    """
+    Represents a call to the .pop() method.
+    
+    Represents a call to the .pop() method which
+    removes the item at the specified index. 
+    The method also returns the removed item.
+
+    Parameters
+    ----------
+    list_variable : TypedAstNode
+        The name of the list.
+
+    index_element : TypedAstNode
+        The current index value for the element to be popped.
+    """
+    __slots__ = ('_dtype','_precision', '_index','_list_variable')
+    _attribute_nodes = ('_index','_list_variable')
+    _rank = 0
+    _order = None
+    _shape = None
+    _class_type = NativeHomogeneousList()
+    name = 'pop'
+
+    def __init__(self, list_variable, index_element=None):
+        self._index = index_element
+        self._list_variable = list_variable
+        self._dtype = list_variable.dtype
+        self._precision = list_variable.precision
+        super().__init__()
+
+    @property
+    def pop_index(self):
+        """
+        The current index value for the element to be popped.
+
+        The current index value for the element to be popped.
+        """
+        return self._index
+
+    @property
+    def list_variable(self):
+        """
+        Provide the name of the list as the return value.
+        
+        Provide the name of the list as the return value.
+        """
+        return self._list_variable
+
+#==============================================================================
+class ListClear(PyccelInternalFunction) :
+    """
+    Represents a call to the .clear() method.
+    
+    Represents a call to the .clear() method which deletes all elements from a list, 
+    effectively turning it into an empty list.
+    Note that the .clear() method doesn't return any value.
+
+    Parameters
+    ----------
+    list_variable : TypedAstNode
+        The name of the list.
+    """
+    __slots__ = ('_list_variable',)
+    _attribute_nodes = ('_list_variable',)
+    _dtype = NativeVoid()
+    _precision = -1
+    _rank = 0
+    _order = None
+    _shape = None
+    _class_type = NativeHomogeneousList()
+    name = 'clear'
+
+    def __init__(self, list_variable):
+        self._list_variable = list_variable
+        super().__init__()
+
+    @property
+    def list_variable(self):
+        """
+        Provide the name of the list as the return value.
+
+        Provide the name of the list as the return value.
+        """
+        return self._list_variable