+++ pyccel/ast/class_defs.py
@@ -6,7 +6,7 @@
 This module contains all types which define a python class which is automatically recognised by pyccel
 """
 
-from pyccel.ast.builtin_methods.list_methods import ListAppend
+from pyccel.ast.builtin_methods.list_methods import ListAppend, ListPop, ListClear
 
 
 from .builtins  import PythonImag, PythonReal, PythonConjugate
@@ -138,6 +138,10 @@ StringClass = ClassDef('string', class_type = NativeString(),
 ListClass = ClassDef('list', class_type = NativeHomogeneousList(),
         methods=[
             PyccelFunctionDef('append', func_class = ListAppend,
+                decorators = {}),
+            PyccelFunctionDef('pop', func_class = ListPop,
+                decorators = {}),
+            PyccelFunctionDef('clear', func_class = ListClear,
                 decorators = {}),
         ])